requirements.txt
--index-url https://download.pytorch.org/whl/cu124
transformers[torch]==4.46.3
torch==2.5.1
webvtt-py==0.5.1
datasets==3.1.0
beautifulsoup4==4.12.3
pycryptodome==3.21.0
nltk==3.9.1
lxml==5.3.0